# docs/api-servas-swagger.yaml
openapi: 3.2.0
info:
  title: API Reserva de Servicios
  description: Especificación alineada al plan de trabajo oficial y las Historias de Usuario (HU-001 a HU-040).
  version: 1.0.0
paths:
...
  /companies:
    post:
      tags: [Compañías & Negocios]
      summary: Registrar negocio (HU-003)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [nit, name, address]
              properties:
                nit: { type: string }
                name: { type: string }
                description: { type: string }
                address: { type: string }
                social_media: { type: string }
                logo: { type: string, format: binary }
      responses:
        '201': { description: Compañía creada. }
  /companies/{id}:
    put:
      tags: [Compañías & Negocios]
      summary: Editar datos del negocio (HU-003)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                address: { type: string }
                social_media: { type: string }
                logo: { type: string, format: binary }
      responses:
        '200': { description: Negocio actualizado. }

  # ==========================================
  # ÉPICA 2: CATÁLOGO DE SERVICIOS
  # ==========================================
...
  /services:
    get:
      tags: [Maestros & Catálogo]
      summary: Consultar catálogo público de servicios (HU-007, HU-008)
      parameters:
        - name: comuna_id
          in: query
          schema: { type: integer }
        - name: modality
          in: query
          schema: { type: string, enum: [VIRTUAL, PRESENCIAL] }
        - name: company_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Lista de servicios disponibles con costo, modalidad, marca y comuna. }

    post:
      tags: [Maestros & Catálogo]
      summary: Registrar servicio (HU-009, HU-013)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_id, comuna_id, name, modality, cost, duration_minutes]
              properties:
                company_id: { type: string, format: uuid }
                comuna_id: { type: integer }
                name: { type: string }
                modality: { type: string, enum: [VIRTUAL, PRESENCIAL] }
                cost: { type: number }
                duration_minutes: { type: integer }
                description: { type: string }
                recommendations: { type: string }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
      responses:
        '201': { description: Servicio creado. }
  /services/{id}:
    get:
      tags: [Maestros & Catálogo]
      summary: Consultar detalle del servicio (HU-008, HU-009)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Detalle completo con recomendaciones, modalidad y marca. }

    put:
      tags: [Maestros & Catálogo]
      summary: Modificar información de un servicio (HU-011)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comuna_id: { type: integer, description: "Obligatorio si la modalidad es PRESENCIAL." }
                name: { type: string }
                modality: { type: string, enum: [VIRTUAL, PRESENCIAL] }
                cost: { type: number }
                duration_minutes: { type: integer }
                description: { type: string }
                recommendations: { type: string }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                is_active: { type: boolean }
      responses:
        '200': { description: Servicio actualizado. }
  /services/{id}/status:
    patch:
      tags: [Maestros & Catálogo]
      summary: Activar o desactivar servicio (HU-012)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active: { type: boolean }
      responses:
        '200': { description: Estado del servicio actualizado. }

  # ==========================================
  # ÉPICA 3: GESTIÓN DE AGENDAS Y DISPONIBILIDAD
  # ==========================================
...
  /reports/reservations:
    get:
      tags: [Reportes]
      summary: Reporte de reservas realizadas (HU-038)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Métricas agregadas de reservas efectuadas. }
  /reports/occupancy:
    get:
      tags: [Reportes]
      summary: Nivel de ocupación por horarios (HU-039)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Porcentaje de uso sobre la capacidad configurada. }
  /reports/demand:
    get:
      tags: [Reportes]
      summary: Servicios de mayor demanda (HU-040)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Ranking de servicios más solicitados. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT